updated a little

+++ Backend/app/models/department.py
@@ -4,7 +4,7 @@ Role: Model layer. It maps Python objects to database tables and relationships.
 """
 
 # app/models/department.py
-from sqlalchemy import Column, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, Column, ForeignKey, String, UniqueConstraint
 from sqlalchemy.orm import relationship
 from app.models.base import Base
 from app.models.associations import program_department_association, event_department_association
@@ -15,8 +15,8 @@ class Department(Base):
         UniqueConstraint("school_id", "name", name="uq_departments_school_name"),
     )
 
-    id = Column(Integer, primary_key=True, index=True)
-    school_id = Column(Integer, ForeignKey("schools.id", ondelete="CASCADE"), index=True, nullable=True)
+    id = Column(BigInteger, primary_key=True, index=True)
+    school_id = Column(BigInteger, ForeignKey("schools.id", ondelete="CASCADE"), index=True, nullable=True)
     name = Column(String, nullable=False)
 
     # Relationships
