+++ Backend/app/models/program.py
@@ -4,7 +4,7 @@ Role: Model layer. It maps Python objects to database tables and relationships.
 """
 
 # app/models/program.py
-from sqlalchemy import Column, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, Column, ForeignKey, String, UniqueConstraint
 from sqlalchemy.orm import relationship
 from app.models.base import Base
 from app.models.associations import program_department_association, event_program_association
@@ -15,8 +15,8 @@ class Program(Base):
         UniqueConstraint("school_id", "name", name="uq_programs_school_name"),
     )
 
-    id = Column(Integer, primary_key=True, index=True)
-    school_id = Column(Integer, ForeignKey("schools.id", ondelete="CASCADE"), index=True, nullable=True)
+    id = Column(BigInteger, primary_key=True, index=True)
+    school_id = Column(BigInteger, ForeignKey("schools.id", ondelete="CASCADE"), index=True, nullable=True)
     name = Column(String, nullable=False)
 
     # Relationships
